vm1/fastapi/requirements.txt
fastapi
uvicorn
kafka-python
sqlalchemy
asyncpg